Update README.md
his is a full application build of the story writing app which uses Node Js, Express Js, MongoDB, Passport Js with a Google OAuth strategy, and more. In designing part with the normal HTML CSS I also used material UI for better quality. Handlebars with different types of helpers are used as the templating language